+++ src/main/java/edu/infsci2560/Storage/StorageException.java
@@ -1,0 +1,12 @@
+package edu.infsci2560.Storage;
+
+public class StorageException extends RuntimeException {
+
+    public StorageException(String message) {
+        super(message);
+    }
+
+    public StorageException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}