+++ src/main/java/edu/infsci2560/Storage/StorageFileNotFoundException.java
@@ -1,0 +1,11 @@
+package edu.infsci2560.Storage;
+public class StorageFileNotFoundException extends StorageException {
+
+    public StorageFileNotFoundException(String message) {
+        super(message);
+    }
+
+    public StorageFileNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}