+++ src/main/java/edu/infsci2560/repositories/ActorRepository.java
@@ -1,0 +1,6 @@
+package edu.infsci2560.repositories;
+import edu.infsci2560.models.Actor;
+import org.springframework.data.repository.PagingAndSortingRepository;
+
+
+public interface ActorRepository extends PagingAndSortingRepository<Actor, Long> {}