+++ src/main/java/edu/infsci2560/repositories/MostPopularRepository.java
@@ -1,0 +1,7 @@
+package edu.infsci2560.repositories;
+import edu.infsci2560.models.MostPopular;
+import org.springframework.data.repository.PagingAndSortingRepository;
+
+
+public interface MostPopularRepository extends PagingAndSortingRepository<MostPopular, Long> {}
+