+++ src/main/java/edu/infsci2560/services/ComingService.java
@@ -1,0 +1,49 @@
+package edu.infsci2560.services;
+
+import edu.infsci2560.models.Coming;
+import edu.infsci2560.models.Coming.MovieType;
+import edu.infsci2560.repositories.ComingRepository;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+/**
+ *
+ * @author Hannah
+ */
+@RestController
+@RequestMapping("/public/api/update")
+public class ComingService {
+
+    @Autowired
+    private ComingRepository repository;
+
+    @RequestMapping(method = RequestMethod.GET, produces = "application/json")
+    public ResponseEntity<Iterable<Coming>> list() {
+        HttpHeaders headers = new HttpHeaders();
+        return new ResponseEntity<>(repository.findAll(), headers, HttpStatus.OK);
+    }
+
+    @RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = "application/json")
+    public ResponseEntity<Coming> list(@PathVariable("id") Long id) {
+        HttpHeaders headers = new HttpHeaders();
+        return new ResponseEntity<>(repository.findOne(id), headers, HttpStatus.OK);
+    }
+
+    @RequestMapping(method = RequestMethod.POST, consumes="application/json", produces = "application/json")
+    public ResponseEntity<Coming> create(@RequestBody Coming update) {
+        HttpHeaders headers = new HttpHeaders();
+        return new ResponseEntity<>(repository.save(update), headers, HttpStatus.OK);
+    }
+}