+++ src/main/java/edu/infsci2560/services/MostPopularService.java
@@ -1,0 +1,49 @@
+package edu.infsci2560.services;
+
+import edu.infsci2560.models.MostPopular;
+import edu.infsci2560.models.MostPopular.MovieType;
+import edu.infsci2560.repositories.MostPopularRepository;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+/**
+ *
+ * @author Hannah
+ */
+@RestController
+@RequestMapping("/public/api/top")
+public class MostPopularService {
+
+    @Autowired
+    private MostPopularRepository repository;
+
+    @RequestMapping(method = RequestMethod.GET, produces = "application/json")
+    public ResponseEntity<Iterable<MostPopular>> list() {
+        HttpHeaders headers = new HttpHeaders();
+        return new ResponseEntity<>(repository.findAll(), headers, HttpStatus.OK);
+    }
+
+    @RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = "application/json")
+    public ResponseEntity<MostPopular> list(@PathVariable("id") Long id) {
+        HttpHeaders headers = new HttpHeaders();
+        return new ResponseEntity<>(repository.findOne(id), headers, HttpStatus.OK);
+    }
+
+    @RequestMapping(method = RequestMethod.POST, consumes="application/json", produces = "application/json")
+    public ResponseEntity<MostPopular> create(@RequestBody MostPopular top) {
+        HttpHeaders headers = new HttpHeaders();
+        return new ResponseEntity<>(repository.save(top), headers, HttpStatus.OK);
+    }
+}